New version of sacamantecas
This is a rewrite of sacamantecas in order to add some new features and
prepare the code to be more robust and flexible.

It is more or less identical in functionality to the previous version
but the output is different and can work in a kind of test mode by using
text files instead of Excel files.

The changes are too many to mention here and this version should in fact
be considered the FIRST version of sacamantecas from now on.