Modify HTML parser, make it a Manteca Skimmer
In order to support URL processing profiles, it is better if the parser
handles everything related with URL processing, including retrieval.

So, now the HTML parser is more than a simple HTML parser, it is a full
featured Manteca Skimmer. It retrieves the URL contents, even performing
redirections if needed, and then it parses the retrieved contents, all
in one go.

To achieve this, the retrieve_uri_contents() code has been incorporated
into the parser, and the entire class has been renamed.