Add support for URL processing profiles
Add support for different profiles for URL processing. This support is
preliminary and it may be too simple for most cases, but works for the
current set of tests.

With the purpose of supporting this, a new configuration file has been
added ('sacamantecas.ini'). It has a simple INI format, with sections
corresponding to the different profiles, each section containing three
keys, to wit: a regex for matching URIs against the profile ('u_match')
and the classes which will be matched to find a metadata key ('k_class')
or a metadata value ('v_class').

All this may change in the future, as this scheme is way too simple for
non-trivial cases, which only can be matched using XPath.

But for now, it works.