Ignore some pylint warnings
Yes, the 'main()' function is too long, but really it is not too complex
and it makes sense to keep the code as-is, without refactoring, because
it is easier to read.

Maybe in the future some code will be refactored into the classes which
already exist, specially if the HTML processing is made more complex and
robust, but for now it works.