Remove last CONFIG_ constant
CONFIG_ global constants does not make any sense anymore, specially this
last one which is a marker to be added in the column heading for all the
new metadata retrieved from the URIs and added back to the Excel file.

Since it is a very particular usage, it is quite internal, it does not
need to be configurable for now, and in the end it can be put back into
a different (and more aptly named) global constant or even in the config
file recently added, it is better to remove it for now.