Change program path detection system
Previously a simple approach using sys.modules['__main__'].__file__ was
being used, and it worked for the most part, in fact it was preferred
over other more complex methods. Thing is, it fails when the script is
'frozen', that is, converted to a standalone executable.

So now the system is simpler, using sys.executable when frozen, __file__
when run as a normal Python script.