Precompile regular expressions
Python already caches regular expressions when compiling them, but the
cache size is limited and when using a lot of regular expressions (that
is, a lot of URL processing profiles), the cache may end up full. Since
there are a high amount of regular expression searches, it is not a bad
idea to precompile them, just in case.