Do not add empty metadata to Skimmed files
Before, empty metadata was being added to the Skimmed files. For Excel
this was not a problem because empty metadata was not added anyway, but
for text files it is a problem. Not a great deal, because adding empty
metadata just meant a bit of noise in the output file, but the URI was
being written to the file and that does not make any sense, having the
URI but not metadata. So, it is better if empty metadata is ignored.