Remove MessageBox() from error()
The MessageBox() existed, mainly, because previously the program did not
output anything to the console, and the logging was a bit limited. That
modal MessageBox was the only way of knowing something bad happened. But
now the program has console output and a detailed logging system which
is much better, so it makes sense to remove that obsolete code, having
less dependencies as a result.