Improve keypress waiting at program end
Before, the program was waiting for a keypress before finishing, just in
case it was launched on a transient console window. This avoided a nasty
problem: the window may close at program end and then the program output
would be lost.

It worked, but only if the program output was normal, from operation. In
case of fatal errors which happened before normal operation started, the
keypress waiting was not happening.

Now, atexit is used to make sure it always happens.