Add setup and building system
In order to automate virtual environment setup and executable building,
a new system for doing both things have been created.

The venv.cmd file creates a virtual environment for this project, and
populates it with the packages specified in requirements.txt file.

The build.cmd file activates the virtual environment (creating it if it
does not exist) and then builds the executable.

The .gitignore file has been modified so its first line is the name of
the virtual environment directory, which makes automation much simpler,
allowing to change that name in a single place.