Create ZIP bundle in building system
Before, the ZIP bundle was created by hand. Now it is created using the
new building system. To wit, using PowerShell and the Compress-Archive
cmdlet, which is present in PowerShell 4.0 and above, so it is safe to
assume it is usable.

An alternative is to use tar in Windows 10 build 17063 or above, but the
PowerShell method works in more systems, apparently, and even though it
is slower the performance is more than acceptable.