Add support for heading-less Excel files
If the input Excel file did not contain a heading row, the program was
confused and tried to write the headings on the first row nonetheless,
wreaking havoc and producing a heading-less Excel with a first row which
looked a bit weird, with heading-colored cells but not heading, having
metadata there instead.

That is now fixed and the program generates its own heading row, even if
a heading row already exists, keeping things separated and clean.