New building system
The previous building system was written using a couple of batch files,
and as such was not specially fast or robust. Probably maintenance would
be hard and readability was suboptimal.

Since the entire project depends on Python, the sensible decision was to
rewrite the building system in Python. This first version may not be the
most robust or readable, but it is better than the batch files.

The new building system creates the virtual environment, populates it,
and then builds the executable and creates the compressed bundle for the
GitHub releases.

The .gitignore file has been affected by this change, of course.