Improve and sanitize Skimmer
In order to support profiles where key or value classes are a subset one
of the other, the HTML parser (the Skimmer) has been slighty improved.

Now, the matches are full string matches by default, which also is more
predictable.

Also, nested key-value tags are now explicitly NOT supported, so parsing
has been a bit sanitized.