Improve handling of multiline log messages
The format for multiline log messages was visually confusing because it
used a multiline marker BEFORE the actual line contents, so breaking the
alignment between lines and looking like garbled characters.

Now, the marker is inserted at the end of continued lines, which is more
or less unobstrusive while still marking the fact that the log message
has multiple lines.