Use custom User-Agent string
Until now the default urllib User-Agent string was being used, and this
was not a great idea for a whole variety of good reasons.

Now, a customized User-Agent string providing the real identity of the
program, the platform details and the URI where the program can be found
is used instead.