Refactor sources processing code
Some code has been moved from the main skimming function to the function
which parses the input arguments, as it was much more related to argv
handling than to the actual processing of the sources specified in the
command line.

So, the constants related to the sources types are no longer needed and
have been removed.