Improve INI file handling
Before, 'Baratz' profiles had to have EMPTY 'k_class' and 'v_class' keys
and this is very inconvenient because in the end that is syntactic noise
within the INI file. Now, 'Baratz' profiles DO NOT have 'k_class' and
'v_class' keys, which is cleaner.

Also, the 'u_match' key has been renamed to 'uri', which is clearer.