Improve parser selection code
Before, the keys defined inside OldRegimeParser were used literally for
comparison against the keys defined by the profile, so see if this was
the needed parser instead of BaratzParser. This was probably not very
fast but also was difficult to read and tighly coupled the calling code
with the parser code, unnecessarily because the calling code did not use
those keys at all, except for discriminating.

Now, this is done programatically, checking all defined parsers to find
one that can parse the profile (the first matching one is used).

Also, a bit of parser code cleanup was done in order to support this.