Add 'dump mode' for Manteca sources
From now on, if a Manteca source starts with 'dump://', its URIs are not
processed for metadata. Instead, all URI contents are dumped into HTML
files so they can be reused later (e.g. for testing) without the need of
actually retrieving them from the network.

This means that the testing can be done faster, without Internet access
and without repeteadly hitting servers.