Improve regex matching code in Old Regime parser
Before, a fullmatch() was being used, but unfortunately the base HTML
parser used did not split tag attribute values into a list, and a full
match was failing in cases where the value was being found but was not
the ONLY value. A search() is done now, instead, which can lead to false
positives but that can be fixed by fine-tuning the corresponding regular
expressions in the config file.