Improve HTML nesting handling in Old Regime parser
The parser is not stricter when dealing with HTML nesting.

Before, the closing tags were not handled, and the parser was a bit too
much lenient with HTML nesting, considering ANY closing tag as the end
of the current parsing block (keys or values). This caused problems for
multiple values, nested values, etc. while at the same time allowing for
VERY ill-formed HTML code when in reality catalogues have weird but not
ill-formed HTML.