Allow empty keys, multivalues in Old Regime parser
Before, multiple values where not allowed if they apperared in DIFFERENT
HTML elements. Now they are allowed and presented as a single value with
the different values found separated by a marker.

Also, before, empty keys were considered incomplete metadata and as so,
invalid. The corresponding values were ignored. Now they are presented
under a specially named key.

Last but not least, the parser is now more permissive when dealing with
missing keys when multiple values are hierarchically under a single key.
In that case, only the first value had a key before, while the rest were
handled as if the key was empty. Now the parser considers that the key
for those values is the same as for the first one found.