Pause at the end only in real consoles
Until now, the message at the end of the program execution warning the
user to press any key to finish was shown unconditionally. The purpose
was to avoid any command prompt window to close automatically when the
program finished running, thus losing program output.

The problem was that the message was being shown EVEN if the output of
the program was redirected to a file or NUL. Now, the message is shown
only if the output is a real console.