Improve virtual environment detection
Before, the existence and proper set up of the virtual environment was
checked every time it was needed to use it. But this is overkill, since
sometimes expensive checks can be avoided if the virtual environment is
activated.

So, that simple and cheap check is done in some places so the expensive
test can be done only when needed.