Add exit code to sacamantecas.py
Previously the return code was always 0, indicating success. That was so
even if wrong happened during execution because the program tried hard
to keep executing after recoverable errors.

Now, if anything wrong happens during execution, the program returns a
nonzero status indicating that, even if some parts of the execution were
successful.