Fix output encoding problem
Under Windows, sys.stdout and sys.stderr encoding is not the same when
they are redirected to a file or captured by subprocess.run(). So, the
encoding has been explicitly set on both streams so they use the same
no matter if outputting to a console or being redirected to a file.