Rewrite entire test suite
The automated unit testing system has been rewritten so it is easier to
expand in the future. Now it consists on a set of classes and no just a
couple of ad-hoc functions.

The changes are too many to mention here as it is a full rewrite rather
than a refactor.

Python unittest has not been used because it was cumbersome to use, and
quite difficult to customize. This code works better for the time being.