Refactor code in make.py
The next step for integrating all helper scripts is to reorganize code
in make.py so different 'make targets' can be run, not only building the
frozen executable.

Also, a basic framework to handle the command line so all the different
targets can be invoked with different arguments on the command line has
been added.