Add a global configuration object to make.py
In order to simplify the code, and to better structure target handlers,
to make easier adding target handlers in the future, a global object to
store the script configuration has been added.

This way, target handlers do not need parameters, and more handlers can
be added in the future in a much easier manner.