Remove banner comments
This is the first time I use the first person for a commit, but I think
it is necessary. Until now I was using a minimap in Visual Studio Code,
and in order to find functions fast I was using banner comments, which
would stand out on the minimap.

This had a few problems: it needed an extension, that extension was not
fast to use because I never bothered to add a keyboard shortcut, a LOT
of lines were added to the source code and in some occasions the text of
the comment was too long to fit once the banner font was used.

But those problems were not that important. The REAL problem was that I
was using the banner comments and the minimap instead of the proper tool
for the tast: a code navigation tool. For a million reasons.

Any decent editor MUST provide such code navigation tool, using banner
comments and a minimap to navigate code is subpar, period.

So, banner comments have been removed from the code.