Add basic logging support when code is imported
When the code is imported the logging system cannot work the same as it
does when the code is run, since no logging files should be created, the
formatting should be different, etc.

Since the code is imported for testing and for some unit tests it may be
necessary to examine logging output, a very basic logging configuration
is now enabled when the code is imported.