Disable waiting for keypress when code is imported
The program was waiting for a keypress when the console was detected as
transient even when the code was imported by the unit testing framework.

Since the unit testing framework may change in the future, adding checks
in the waiting for keypress code to detect if the code is being run by
the unit testing framework, it is better to never wait for a keypress at
program termination if the code is imported instead of being run.

So, the code to handle whether the program should wait for a keypress at
termination has been moved to the area which is only run if the program
is not imported.