Rewrite command-line arguments processing
In order to make the code easier for unit testing, the function that did
the command-line arguments processing has been modified so it is simpler
and has less responsibilities.

Also, the name has been changed from process_argv() to parse_sources(),
since now the function is more generic and can process the command-line
arguments if sys.argv is provided, or any other list containing sources
for data retrieval and processing.