Decorate main() for handling KeyboardInterrupt
The main goal is to extract KeyboardInterrupt handling from main(), so
the function has simpler logic, more explicit and simpler return points,
and less useless indentation, which are problems caused by the fact that
the try..except block has to entirely surround the main function for the
handling to be full and precise.

So, that handling has to be extracted to an external function or to the
top-level code which calls main. But in the second case, the function is
not going to have KeyboardInterrupt handling when the code is imported,
or under unit testing, and this may be undesirable.

In short, the only solution is handling KeyboardInterrupt in an external
function which in turn calls main(), and then the best solution is using
a decorator because it is more elegant, more concise and more flexible.