Decorate main() for logging
The main goal is to extract logging setup/teardown from main(), so the
function has simpler return points and the logic for logging setup and
teardown is completely separated from the rest of business logic.

So, logging setup/teardown hase to be extracted to top-level code which
calls main or to an external function. But in the first case, loggign is
not going to work when the code is imported or under unit testing, and
this may be undesirable.

In short, the only solution is to implement this setup and teardown in a
function which in turn calls main(), and then the best solution is using
a decorator because it is more elegant, more concise and more flexible.