Make error() identical to logging.error()
The main reason the error() function existed was to set the global error
flag and output an error preamble to sys.stderr.

The error flag was removed a long time ago and with the improved logging
setup the preamble can be moved into a logging formatter, which is where
it should be. And it has been moved there.

So, now the error() function is identical to logging.error(). After the
messages are put into constants, the function will be removed. Right now
it is just deprecated.