Take back error() and warning() helper functions
Turns out that using logging formatters to prepend a header for some of
the messages, depending on the logging level, was a BAD IDEA. It works,
and it does not prevent any fine-grained control needed for formatting,
but makes impossible to use one logging handler per logging destination.

This is due to the fact that a handler can use only one formatter, so if
different formats are needed for the same logging level depending on the
logging destination, more handlers have to be created.

So the helpers are back and the logging system uses one handler for each
destination, with a formatter depending ONLY on the destination.