Move private test fixture to where it is used
The 'unreadable_file' test fixture was defined in conftest.py but it was
used by test_load_profiles.py only. The reason behind this decision was
to avoid a linting error.

This should be fixed properly and not by putting a private test fixture
in the global test configuration file.