Remove support for dumping sources
At least, remove support for dumping sources in the way it existed until
now, because it was only used for generating testing files. The reason
behind this was that by dumping some sources into files the application
could be tested for regressions without having to actually contact any
server. This was very useful with the previous testing system, but now
that all testing is done using unit testing, it no longer makes sense.