Rewrite source handlers completely
The previous design, which was being refactored, consisted on a bunch of
very simple classes, all of them providing a single method, which led to
a lot of boilerplate code and a messy design.

The new design are just functions, more or less the same single methods
that the original classes provided, but with a simpler code and a code
much easier to follow.

These functions are implemented as generators.

Corresponding tests have been modified.