Convert handlers to the new URL handling system
In addition to this, and because it makes sense to have these changes on
a single commit, the handler for spreadsheets has been fixed. The code
was just copypasted from the original code, and it did not work when all
code is in the same function instead of being separated in two classes.

It is now fixed, too.