Filter URLs in the handlers
Until now, ALL lines were considered potential URLs, and the filtering
was done by the parser, URL retriever, etc. but not by the URL provider.

It makes sense to do this in the handlers. The spreadsheet file handler
was already doing it, and the single URL handler was already receiving
a proper URL, so explicit testing was in order.