Refactor application messages.
For a start, they are now stored in a SimpleNamespace, not an StrEnum.

Also, they have been reordered, some of them have been renamed and some
isolated string constants have been moved to the namespace, too.

In addition to this, some messages have been moved around in the code.

Finally unit tests have been modified to reflect this change.