Perform platform test earlier
Since the application only works on Win32, the platform test should be
done as soon as possible.

It was being delayed because it used a message string whose content had
to be computed before performing the test. This is no longer needed.

The only drawback is that the error message is hardcoded rather than be
put in a string constant, but this is a very minor issue since this test
will only output an error message if the application is run on the wrong
platform, and that is not going to happen frequently.