Replace SimpleNamespace with StrEnum for messages
Although a SimpleNamespace object looks better for storing application
messages, the attributes (in this case, the identifiers for the strings)
are dynamically created at runtime and that causes problems with IDEs.

So, in order to have refactoring and completion help from the IDE, using
a StrEnum is much better and in the end, conceptually, it is similar.