Remove custom exception in parse_sources()
Before, the function raised a custom exception for unsupported sources,
but the function is really a generator, so after finding the first not
supported source, the generator stopped.

This was the intended behavior but now the preferred mode of operation
is just notifying about invalid sources but keep processing the rest, if
any, since it is perfectly possible for the application.

The main() code and unit tests have been updated accordingly.