Reorder functions in all source files
Even though current ordering usually makes sense, it is better to apply
a consistent style for the order of functions in source files.

From now on they will be sorted according to the order they are called
in the code, with global helpers first, except for main() which will be
at the very end. This way the entire file more or less resembles main().

For unit tests, functions will be ordered just like they are in the code
under test if possible, or a custom order depending on the unit test.