Rewrite single URL handler
For a start, the output file was not having the proper stem marker, so
some wrong error message was returned in bootstrap(). No big deal, but.

Second, but more important, the handler was notifying about successful
initialization when, in fact, it was not initialized, since the output
file creation was being delayed until the handler actually got metadata
instead of being created at initialization time.

Both problems are now solved.