Simplify custom exceptions
The current custom exceptions do not have any custom argument, they are
just aliases for Exception to improve code readability. But previously
they were getting a 'details' argument containing the error message.

That is handled automatically by Exception itself, so there is no need
to implement it, and this leads to cleaner code where the exception is
handled or used.

Unit tests have been modified accordingly.