Adopt semantic versioning for the application
The application version number almost followed the semver 2.0.0, lacking
the minor number only, since development versions HAD a pre-release tag
and build metadata is optional.

The minor version number was not included because the patch number was
abused to indicate both backward compatible changes and bug fixes.

This is not the end of the world, of course, but semantic versioning is
much more informative, complete and better structured, so adopting it is
the way to go.

The make script has been adapted to use this new versioning approach.