Fix string conversion for error() arguments
The error() function was dealing with arguments in a way that arguments
were converted to strings implicitly. But with the latest change in the
logging system, explicit conversion was needed for object arguments.

Instead of requesting this conversion when calling the function, it will
happen inside the function itself, which is more transparent.