Separate exception messages from details
Now the entire custom exception hierarchy explicitly separates the error
message from the error details (that is, optional additional information
which may be useful when reporting the error) in the constructors.

For that, the exception hierarchy has been modified, and now all custom
exceptions support the 'detail' argument, not only SkimmingError, and of
course the attribute has been normalized, because ProfilesError used the
'reason' attribute for additional information in some contexts.

Unit tests have been adapted to the new system.