Convert profiles from dictionaries to namedtuples
Until now profiles were stored in dictionaries, but since the main keys
are fixed (the URL pattern, the corresponding parser, the parser config,
etc.) it makes sense to treat a profile like a 'data object' with fixed
keys, that is, with attributes. So, a namedtuple.

The load_profiles() function and the corresponding unit tests have been
updated to adapt to this new data pattern.