Improve constants
They have been moved up in the file, as much as possible, so ALL strings
in the application can potentially be added as a constant, and constants
are used as much as possible instead of magic numbers or other hacks.

In addition to this some of the constants have been renamed or improved.