Move some constants where they are used
Some of the constants are used by just one function, most of the time as
a means of avoiding typing errors, which is typical with string literals
and using a constant is a very good idea even though it is used once.

So, it makes sense to move the constants from the beginning of the file
to where they are used, at least for now, as they are not configuration
constants.

This means that some constants have been brought back, others have been
moved and others are kept at the top of the file even if they are used
just once because they are considered configuration constants.