Optimize exception hook
Some of the private string constants have been removed after the string
formatting has been simplified. Now the strings are more readable.

Also, the output message from the exception hook is now less informative
for the end user but MORE informative for the developer and even allows
for some automation. The output message is not useful for end users, the
line numbers, codes, introspection, etc. is useless unless familiar with
the source code, and the end user should not be familiar with it.

So the information presented has to be useful for developers, not users.