Improve initialization of parsers
Instead of creating the needed attributes on __init__() and assign them
proper initialization values, definition and initialization are now two
separate actions. On __init__() the attributes are defined but all their
values will be None. On reset(), proper initialization values are used
for the attributes.

This makes easier to spot if some attribute was properly initialized or
not, which is never going to happen but… para bellum.