Revert some of the constants
The 'all strings will be constants' experiment has been quite a success,
*except* for the fact that has complicated code outline by a factor of 5
and has made some parts of the code unnecessarily verbose.

Still, the experiment has been a great idea because it showed that the
code was sprinkled with string literals that could be refactored making
the code safer (no as many undetected typos) and more readable.

The problem is that refactoring was done in an all-or-nothing fashion,
and nearly ALL string literals were put into constants, even when it did
not improve code safety or readability.

In addition to this, some 'local' constants have polluted the namespace
a bit, making code actually worse.

So, some of the constants will be removed, some will be kept, and others
will be moved into a new 'Config' namespace, etc.

Unit tests have been modified accordingly.