Fix logging fallback configuration bug
When the logging system cannot be properly initializad a fallback config
was used. But the code was anyway trying to use the logging indentation
helpers, which are not available unless the logging system is configured
in full.

Now, sane placeholders are used for this case, so the code can use the
indentation helpers even though the logging system is not initialized
properly.