Hardcode virtual environment path where needed
The virtual environment path was being obtained from the contents of the
.gitignore file. This way the path could be changed at will and affected
code would use the new value.

This was flexible but overkill, complicating the code unnecessarily, and
what is worse, making automatic virtual environment creation (e.g. using
IDE facilities, etc.) more complex and error prone.

In short, it is not worth the pain, and in order to simplify make.py, it
is much better to have this value hardcoded. The chosen value is the one
used by default by all Python tools, the '.venv' subdirectory.