Move DEVELOPMENT_MODE constant to version.py
Since the concept of 'development mode' depends heavily on the version
code, to wit, if it denotes a prerelease version, it is only logical to
move that constant definition into version.py rather than implementing
the logic in the main script using V_PRERELEASE from version.py file.

The result is exactly the same, but the code is more readable.