Return detected encoding/charset in retrieve_url()
The function gets the URL contents as a byte string but until now it was
returning a Unicode string, decoded using the charset/encoding detected
from the URL itself. Now, it returns both the original byte string and
the encoding in a tuple, allowing for more flexibility when the contents
are handled by the caller.

Also, the fallback charset constant has been given a better name.

Affected unit tests have been fixed.

Even though HTML uses 'charset' as a synonym for 'encoding' the function
uses 'charset' internally (because it is 'speaking' with HTML servers)
but 'encoding' externally, so it returns an encoding, not a charset.