Move APP_NAME into main script
It was moved to version.py so it could be imported in unit tests, or in
other helper scripts (like the building script), but it really did not
belong there.

So, it is now defined within the main script, and properly computed from
the script name itself, rather than being hardcoded, even though it is
not very probable it will change in the future.

Accessing APP_NAME from other scripts requires importing the main script
instead of version.py, but this is now much safer since the side-effects
have been reduced as much as possible.

Affected scripts have been updated.