Perform platform detection much earlier
Now it is done first thing in the main script. The drawback is that the
test cannot use the script constants or other facilities, and to avoid
problems the error message has to be terse and in English to avoid any
encoding problems. And of course, pylint complains about the imports,
since they are no longer at the top of the file.

Still, it is a better idea because that way there will not be any weird
errors happening due to imports not being available in any unsupported
platforms, slightly older Python versions, etc.