Improve shebang line in main script
The version indicator has been removed. It was there for platforms where
the default provided Python version was 2. But those platforms are not
supported by the script, so it was mostly useless.

Also, providing a version indicator was creating a problem when running
the application under a virtual environment, because the Python version
was not handled correctly by the virtual environment.