Fix stupid indentation bug in the error() function
Indentation for the second line of the full error message was working OK
because the error header started with a newline character, so the value
calculated for the indentation was NOT off by one. But this is an error
and depends on never removing that leading newline from the error header
so a new calculation is used, not taking into account leading whitespace
and using the explicit '+1' needed to properly align the messages.