Improve logging
The logging mechanism was modifying the logging message by stripping it.
This is not what the logging module does, it honors leading and trailing
whitespace, newlines, etc. This is a problem for testing, mocking, etc.
and a potential source of unexpected output, so it has been changed.

Now, the logging mechanism is compatible with the logging module.

In addition to this, some refactoring has been done in the error() and
logging functions so the code is easier to test.

Affected unit tests have been modified.