Simplify determination of the root path
Previously, the root path was determined before the application main()
was run, because the checks were very paranoid and in fact the execution
would be aborted if determination failed.

This abortion would take place if the application cannot be detected as
a frozen executable, and the '__file__' attribute is missing at runtime,
but this is an impossible scenario for this application.

Affected unit tests have been fixed.

So, now the root path is stored at runtime within the Config class, and
assumes that '__file__' exists as a bare minimum.