Improve constants in code
For a start, the Config class, which used to store the so-called program
configuration, has been renamed to Constants, which is more appropriate
because it really stores miscellaneous constants.

Second, the few constants that were defined outside this class have been
moved into the class, unifying the constants handling.

Last but not least, the Config (now Constants) class had a dependency on
the Messages class, and it has been removed.

Affected unit tests and helper scripts have been adapted to this.