Remove EMPTY_STRING constant
The goal of such a constant was, mainly, prevent mistyping a string with
a space rather than an empty string. Also, in certain contexts, it could
improve readability by showing intent rather arbitrariness. So, it was
heavily used throughout the code.

The problem is that even though using the constant could show intent, it
did not improve readability much, since the end result looked messy. The
code was not more readable, it just looked denser.

So, the constant has been removed, using the string literal or another,
more meaningful constant, according to the context.

Affected unit tests have been adapted.