Rename constant for default values in parsers
Previously the same constant was used for keys and values, since default
value for both is actually the same, an empty string, but the resulting
code is nearly unreadable and makes much more sense to separate them, so
in the future a different default value may be used by future parsers.

Affected unit tests have been modified.