Improve unit test for parsing random strings
The amount of random data fed was too high, on average the test ran for
about 50k loops. Errors caused by random data crashing the parser should
be revealed with less data. Of course, more random data gives additional
confidence, but with diminishing returns.

Instead of feeding 1.048.576 random characters, a total of 1024 strings
are fed to the parsers, and each one is fed 10 times but with different
parser state. That should give enough confidence.