Improve handlers return values
Until now, and depending on the phase the handler generator was in, the
return value could be True (after successful initialization), None (for
just advancing the iterations over the generator) or URL strings.

Since the only return value actually used are the URL strings, all the
handlers return strings in every phase. After successful initialization
a constant string is used, and for advancing the iterations the URL is
returned again because is as good as any other value and can come handy
in the future.

An abstract type has been added for the handlers to reflect this change.