Add type annotations to main code
This change does not just imply adding the type annotations themselves,
it also implies some minor code modifications and adjustments because by
adding the annotations some potential (but quite improbable) bugs have
been fixed when appropriate.

The amount of affected code is huge because type annotations end up in
nearly every function, but the changes are actually very minor.

There are a few type annotation problems still unfixed, because they are
related to the logging system, which is going to be replaced.