Replace logging framework
Until now, sacamantecas used the root logger. There is nothing extremely
wrong with it, since this is not a module and side effects should not be
a problem, but it is a bad practice.

In addition to this, the support for logging indentation was implemented
using a questionable hack.

Now, a custom logger class is used instead, and a custom logger instance
is used rather than the root logger.

Both main code and test suite has been adapted to the new system.
but