Remove paranoid type check for spreadsheets
Technically, the input spreadsheet could be in write-only mode, and the
output spreadsheet could be in read-only mode, by sheer bad luck, but in
a normal run this is impossible, because those incompatible modes should
be activated WITHIN the program itself, and it is not doing that.

So, the checks have been removed and the related strings too.